Initial work on CColorManager

+++ src/game/CScout.h
@@ -33,7 +33,7 @@ public:
     short action;
     Boolean nextRotateFlag;
 
-    virtual void IScout(CAbstractPlayer *thePlayer, short theTeam, long longTeamColor);
+    virtual void IScout(CAbstractPlayer *thePlayer, short theTeam, uint32_t longTeamColor);
     virtual void PlaceParts();
     virtual void FrameAction();
     virtual void ControlViewPoint();
